+++ CHANGELOG.md
@@ -99,6 +99,8 @@
 - `network::protocol::codec::encode_auth` now supports already implemented
   authentication via MD5.
 - Bug in `Tuple::format` which could result in undefined behaviour (aka use after free).
+- `tlua::Push` and `tlua::PushInto` can now be used on non-empty structs and enums
+  with default values for constant generics.
 
 ### Deprecated